+++ src/Vector/ASC/Most150PacketFragment.cpp
@@ -74,7 +74,7 @@ Most150PacketFragment * Most150PacketFragment::parse(File & file, std::string & 
     std::smatch match;
     if (std::regex_match(line, match, regex)) {
         Most150PacketFragment * most150PacketFragment = new Most150PacketFragment;
-        most150PacketFragment->time = std::stof(match[1]);
+        most150PacketFragment->time = std::stod(match[1]);
         most150PacketFragment->channel = std::stoul(match[2]);
         most150PacketFragment->frgMask = std::stoul(match[3], nullptr, 16);
         most150PacketFragment->sourceAdr = std::stoul(match[4], nullptr, 16);
